requirements.txt
pathlib2==2.3.5
pytesseract==0.3.4
pytz==2019.3
psutil==5.7.0
tensorflow_object_detection_api==0.1.1
opencv_python==4.1.0.25
Cython==0.29.15
pytest==5.4.1
lxml==4.5.0
scipy==1.4.1
cryptography==2.8
tensorflow==1.15.0
nose==1.3.7
matplotlib==3.1.3
genapi==0.0.8
ipaddr==2.2.0
object_detection==0.0.3
ordereddict==1.1
pickle5==0.0.9
Pillow==7.1.2
protobuf==3.11.3
pyOpenSSL==19.1.0
python_dateutil==2.8.1
wincertstore==0.2
